+++ src/main/java/day1/Day1.java
@@ -8,11 +8,11 @@ import static main.java.FileProcessorUtil.processFile;
 public class Day1 {
     private static final String INPUT_FILE = "src/main/resources/day1/input.txt";
 
-    public static int phase1() {
+    public static long phase1() {
         return processFile(Phase1Parser::parseNumbersFromText, INPUT_FILE);
     }
 
-    public static int phase2() {
+    public static long phase2() {
         return processFile(NumberParser::parseNumbersAndWordNumbersFromText, INPUT_FILE);
     }
 }
