+++ src/main/java/day3/Day3.java
@@ -7,12 +7,12 @@ public class Day3 {
     private static final String INPUT_FILE = "src/main/resources/day3/input.txt";
     // Use the example_input to check based on the problem description example
     private static final String EXAMPLE_INPUT_FILE = "src/main/resources/day3/example_input.txt";
-    public static int phase1() {
+    public static long phase1() {
         Phase1 phase1 = new Phase1();
         return processFile(phase1::parseEngineDiagram, INPUT_FILE);
     }
 
-    public static int phase2() {
+    public static long phase2() {
         Phase2 phase2 = new Phase2();
         return processFile(phase2::findGears, INPUT_FILE);
     }
