+++ src/main/java/day4/Day4.java
@@ -8,11 +8,11 @@ public class Day4 {
     private static final String INPUT_FILE = "src/main/resources/day4/input.txt";
     // Use the example_input to check based on the problem description example
     private static final String EXAMPLE_INPUT_FILE = "src/main/resources/day4/example_input.txt";
-    public static int phase1() {
+    public static long phase1() {
         return processFile(Phase1::calculateWinnings, INPUT_FILE);
     }
 
-    public static int phase2() {
+    public static long phase2() {
         Phase2 phase2 = new Phase2();
         return processFile(phase2::calculateNumberScratchcards, INPUT_FILE);
     }
